closer to having basic tests working...
broke datastore out into binary file operations

made the binary file optional for a delete.

+++ web/js/util/Files.ts
@@ -11,6 +11,11 @@ const rimraf = require('rimraf');
 export class Files {
 
     public static removeDirectoryRecursively(path: string) {
+
+        if (! fs.existsSync(path)) {
+            return;
+        }
+
         rimraf.sync(path);
     }
